requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
cbor2==5.6.5
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
diskcache==5.6.3
filelock==3.16.1
flatbuffers==20181003210633
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.6.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-text-splitters==0.3.0
langsmith==0.1.137
libclang==18.1.1
llama_cpp_python==0.3.1
luma.core==2.4.2
luma.oled==3.13.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.10
packaging==24.1
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.3
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyftdi==0.55.4
Pygments==2.18.0
pyserial==3.5
python-dotenv==1.0.1
pyusb==1.2.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
RPi.GPIO==0.7.1
safetensors==0.4.5
six==1.16.0
smbus2==0.5.0
sniffio==1.3.1
spidev==3.6
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
tokenizers==0.20.1
torch==2.5.1
tqdm==4.66.6
transformers==4.46.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.16.0
yarl==1.17.0
